fixing erros while deploying and the image section in home secction
