+++ app/layout.jsx
@@ -1,26 +1,30 @@
-import '@/styles/globals.css';
-import { Inter } from 'next/font/google'
-import Navbar from '@/components/Navbar/Navbar'
-import Footer from '@/components/Footer/Footer'
-import DataProvider from '@/context/DataProvider'
+import "@/styles/globals.css";
+import { Inter } from "next/font/google";
+import Navbar from "@/components/Navbar/Navbar";
+import Footer from "@/components/Footer/Footer";
+import DataProvider from "@/context/DataProvider";
+import { ThemeProvider } from "@/components/theme/theme-provider";
 
-const inter = Inter({ subsets: ['latin'] })
+const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
-  title: 'Informatician',
-  description: 'Unleash an open source book platform where literature meets community, and knowledge knows no bounds. Explore a diverse collection of books and information in the domain of books and reading.',
-}
+  title: "Informatician",
+  description: "Unleash an open source book platform where literature meets community, and knowledge knows no bounds. Explore a diverse collection of books and information in the domain of books and reading.",
+
+};
 
 export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-      <DataProvider>
-        <Navbar />
-        {children}
-        <Footer />
-        </DataProvider>
+        <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+          <DataProvider>
+            <Navbar />
+            {children}
+            <Footer />
+          </DataProvider>
+        </ThemeProvider>
       </body>
     </html>
-  )
+  );
 }
