+++ app/page.jsx
@@ -2,8 +2,7 @@ import Hero from "@/components/Hero/Hero";
 
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-24">
-      <h1 className='text-6xl font-semibold'>Informatician</h1>
+    <main className="flex min-h-screen flex-col items-center justify-between">
         <Hero />
     </main>
   )
