+++ next.config.js
@@ -1,7 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
     experimental:{
-        appDir:true,
         serverComponentsExternalPackages:['mongoose'],
     },
     webpack(config){
