+++ package.json
@@ -15,16 +15,16 @@
     "axios": "^1.4.0",
     "dotenv": "^16.3.1",
     "eslint": "8.44.0",
-    "eslint-config-next": "13.4.7",
+    "eslint-config-next": "^13.4.8",
     "lucide-react": "^0.257.0",
-    "next": "13.4.7",
+    "next": "^13.4.8",
     "next-themes": "^0.2.1",
     "path": "^0.12.7",
     "postcss": "8.4.24",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-alice-carousel": "^2.7.1",
     "react-bootstrap": "^2.8.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.2.0",
     "react-icons": "^4.8.0",
     "react-parallax-tilt": "^1.7.151",
     "react-rating-stars-component": "^2.2.0",
